+++ manager-bundle/phpunit.xml.dist
@@ -21,7 +21,4 @@
   <listeners>
     <listener class="Symfony\Bridge\PhpUnit\SymfonyTestsListener"/>
   </listeners>
-  <extensions>
-    <extension class="Contao\ManagerBundle\Tests\PhpunitExtension\DeprecatedClasses"/>
-  </extensions>
 </phpunit>
