+++ manager-bundle/src/Resources/config/services.yaml
@@ -64,11 +64,6 @@ services:
             - '@kernel'
             - '%kernel.project_dir%'
 
-    contao_manager.security.logout_handler:
-        class: Contao\ManagerBundle\Security\Logout\LogoutHandler
-        arguments:
-            - '@?contao_manager.jwt_manager'
-
     contao_manager.twig.file_extension_filter_iterator:
         class: Contao\ManagerBundle\Twig\FileExtensionFilterIterator
         decorates: twig.template_iterator
