+++ phpunit.xml.dist
@@ -82,8 +82,6 @@
     </listener>
   </listeners>
   <extensions>
-    <extension class="Contao\CoreBundle\Tests\PhpunitExtension\DeprecatedClasses"/>
-    <extension class="Contao\ManagerBundle\Tests\PhpunitExtension\DeprecatedClasses"/>
     <extension class="Contao\TestCase\ClearCachePhpunitExtension"/>
     <extension class="Contao\TestCase\WarnXdebugPhpunitExtension"/>
   </extensions>
